refact(user): separate invitation coordinator
