+++ src/components/invitation/invitation_controller_types.d.ts
@@ -1,0 +1,7 @@
+interface InviteUserParams {
+    username: string;
+}
+
+interface AcceptInvitationParams {
+    password: string;
+}
