+++ src/components/user/user_controller_types.d.ts
@@ -7,11 +7,3 @@ interface ChangePasswordParams {
     oldPassword: string;
     newPassword: string;
 }
-
-interface InviteUserParams {
-    username: string;
-}
-
-interface AcceptInvitationParams {
-    password: string;
-}
